decouple-calculate-fee
